segment Allocation deleted

+++ app/Http/Controllers/API/PurchaseRequestDetailsAPIController.php
@@ -2061,19 +2061,32 @@ class PurchaseRequestDetailsAPIController extends AppBaseController
 
 
     public function removeAllItems($id) {
-        $purchase_request_details = PurchaseRequestDetails::where('purchaseRequestID',$id)->get();
-        foreach($purchase_request_details as $purchase_request_detail) {
-                $data = PulledItemFromMR::where('purcahseRequestID',$id)->where('itemCodeSystem',$purchase_request_detail->itemCode)->first();
-                if(isset($data)) {
-                        $m_id =$data->RequestID;
-                        $request = MaterielRequest::find($m_id);
-                        $request->isSelectedToPR =  false;
-                        $request->save();
-                        $data->delete();
-                }
+        $purchase_request = PurchaseRequest::find($id);
+        if($purchase_request) {
+            $purchase_request_details = PurchaseRequestDetails::where('purchaseRequestID',$id)->get();
+
+            foreach($purchase_request_details as $purchase_request_detail) {
+                    $data = PulledItemFromMR::where('purcahseRequestID',$id)->where('itemCodeSystem',$purchase_request_detail->itemCode)->first();
+                    if(isset($data)) {
+                            $m_id =$data->RequestID;
+                            $request = MaterielRequest::find($m_id);
+                            $request->isSelectedToPR =  false;
+                            $request->save();
+                            $data->delete();
+                    }
+            }
+            PurchaseRequestDetails::where('purchaseRequestID',$id)->delete();
+    
+            SegmentAllocatedItem::where('documentMasterAutoID', $id)
+            ->where('documentSystemID', $purchase_request->documentSystemID)
+            ->delete();
+    
+    
+            return $this->sendResponse([], 'Item Deleted Successfully');
+        }else {
+            return $this->sendError('Purchase Request not found');
         }
-        PurchaseRequestDetails::where('purchaseRequestID',$id)->delete();
-        return $this->sendResponse([], 'Item Deleted Successfully');
+        
     }
 
 }
